// src/app/model/CityTree.ts
export type CityTree = {
  internal_ref: number;
  ref: string;
  location: string;
  address: string;
  lat: number;
  lon: number;
  genus: string;
  species: string;
  variety: string;
  scientific: string;
  common: string;
  height: number;
  crown: number;
  dbh: number;
  planted: number;
  wikipediaPageUrl: string;
  fellingInfo: any;
};